factor out test_util.py
