+++ tools/permission_prompt_test.py
@@ -4,7 +4,8 @@
 import os
 import unittest
 
-from util import DenoTestCase, test_main, tty_capture
+from test_util import DenoTestCase, run_tests
+from util import tty_capture
 
 PERMISSIONS_PROMPT_TEST_TS = "tools/permission_prompt_test.ts"
 
@@ -140,4 +141,4 @@ def permission_prompt_tests():
 
 
 if __name__ == "__main__":
-    test_main()
+    run_tests()
